fix user model and convert MyRequestsView to functional component
